+++ src/main/java/nextstep/subway/entity/group/factory/add/SectionAddAction.java
@@ -1,4 +1,4 @@
-package nextstep.subway.entity.group.factory;
+package nextstep.subway.entity.group.factory.add;
 
 public interface SectionAddAction {
